Now added to all tests

+++ test/unit/test_authorized_keys.py
@@ -3,6 +3,7 @@ Unit tests for _authorized_keys
 """
 
 import os
+import sys
 from os import path
 from unittest import TestCase, main
 
@@ -15,10 +16,15 @@ class TestAuthorizedKey(TestCase):
     """
 
     def setUp(self):
-        os.mkdir(_FILE_NAME.replace("authorized_keys", ""))
+        if not os.path.isdir(_FILE_NAME.replace("authorized_keys", "")):
+            os.mkdir(_FILE_NAME.replace("authorized_keys", ""))
         os.close(os.open(_FILE_NAME, os.O_RDWR | os.O_CREAT))
-        with open(_FILE_NAME, "w+", encoding="utf-8"):
-            print(_FILE_NAME.split("/")[-1])
+        if os.path.isfile(_FILE_NAME):
+            with open(_FILE_NAME, "w+", encoding="utf-8"):
+                print(f'{_FILE_NAME.split("/")[-1]} created')
+        else:
+            print("Something wrong happened")
+            sys.exit(-1)
 
     def tearDown(self):
         if path.exists(_FILE_NAME):
